Tests/LibWeb/Text/expected/intersection-observer-visual-viewport-root.txt

visualViewport: 533.333x400.000
rootBounds: 800.000x600.000
rootBounds match layout viewport: PASS
